Revert "Update AdminLogoutAction.java"
This reverts commit 246115db1db6a10ed243e2de9b061bf29d0fc4ea, reversing
changes made to 8b40e917d54ea815694ac6f4befd5bf202861e78.

+++ Source/src/it/carcheck/model/bean/RegionBean.java
@@ -11,11 +11,11 @@ public class RegionBean {
 	@Table(name = "region")
 	public RegionBean() {}
 
-    public short getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(short id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -28,6 +28,6 @@ public class RegionBean {
     }
     
     @PrimaryKey(option = PrimaryKeyOption.Auto_Increment )
-    private short id;
+    private int id;
     private String name;
 }
